requirements.txt
python-binance==1.0.20
typer==0.12.3
pydantic==2.8.0
python-dotenv==1.0.1

